Move helpers to jbi.services.jira and bugzilla
