+++ jbi/services/jira.py
@@ -5,14 +5,17 @@ with that REST client
 from __future__ import annotations
 
 import concurrent.futures
+import json
 import logging
 from functools import lru_cache
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from atlassian import Jira, errors
 
-from jbi import environment
+from jbi import Operation, environment
+from jbi.models import BugzillaWebhookComment
 
+from . import bugzilla
 from .common import InstrumentedClient, ServiceHealth
 
 if TYPE_CHECKING:
@@ -21,6 +24,9 @@ if TYPE_CHECKING:
 settings = environment.get_settings()
 
 logger = logging.getLogger(__name__)
+
+
+JIRA_DESCRIPTION_CHAR_LIMIT = 32767
 
 
 @lru_cache(maxsize=1)
@@ -137,3 +143,228 @@ def _validate_permissions(all_projects_perms):
             },
         )
     return not misconfigured
+
+
+class JiraCreateError(Exception):
+    """Error raised on Jira issue creation."""
+
+
+def create_jira_issue(context, bug, jira_project_key, sync_whiteboard_labels) -> str:
+    """Create a Jira issue with the specified fields and return its key."""
+    logger.debug(
+        "Create new Jira issue for Bug %s",
+        bug.id,
+        extra=context.dict(),
+    )
+    comment_list = bugzilla.get_client().get_comments(bug.id)
+    description = comment_list[0].text[:JIRA_DESCRIPTION_CHAR_LIMIT]
+    fields: dict[str, Any] = {
+        "summary": bug.summary,
+        "issuetype": {"name": bug.issue_type()},
+        "description": description,
+        "project": {"key": jira_project_key},
+    }
+    if sync_whiteboard_labels:
+        fields["labels"] = bug.get_jira_labels()
+
+    jira_response_create = get_client().create_issue(fields=fields)
+
+    # Jira response can be of the form: List or Dictionary
+    if isinstance(jira_response_create, list):
+        # if a list is returned, get the first item
+        jira_response_create = jira_response_create[0]
+
+    if isinstance(jira_response_create, dict):
+        # if a dict is returned or the first item in a list, confirm there are no errors
+        errs = ",".join(jira_response_create.get("errors", []))
+        msgs = ",".join(jira_response_create.get("errorMessages", []))
+        if errs or msgs:
+            raise JiraCreateError(errs + msgs)
+
+    issue_key: str = jira_response_create.get("key")
+    return issue_key
+
+
+def update_jira_issue(context, bug, issue_key, sync_whiteboard_labels):
+    """Update the fields of an existing Jira issue"""
+    logger.debug(
+        "Update fields of Jira issue %s for Bug %s",
+        issue_key,
+        bug.id,
+        extra=context.dict(),
+    )
+    fields = {
+        "summary": bug.summary,
+    }
+    if sync_whiteboard_labels:
+        fields["labels"] = bug.get_jira_labels()
+
+    jira_response_update = get_client().update_issue_field(key=issue_key, fields=fields)
+    return jira_response_update
+
+
+def add_jira_comment(
+    context, issue_key, commenter: str, comment: BugzillaWebhookComment
+):
+    """Publish a comment on the specified Jira issue"""
+    formatted_comment = f"*({commenter})* commented: \n{{quote}}{comment.body}{{quote}}"
+    jira_response = get_client().issue_add_comment(
+        issue_key=issue_key,
+        comment=formatted_comment,
+    )
+    logger.debug(
+        "User comment added to Jira issue %s",
+        issue_key,
+        extra=context.dict(),
+    )
+    return jira_response
+
+
+def add_jira_comments_for_changes(log_context, event, bug, linked_issue_key):
+    """Add comments on the specified Jira issue for each change of the event"""
+    comments: list = []
+    user = event.user.login if event.user else "unknown"
+    for change in event.changes or []:
+        if change.field in ["status", "resolution"]:
+            comments.append(
+                {
+                    "modified by": user,
+                    "resolution": bug.resolution,
+                    "status": bug.status,
+                }
+            )
+        if change.field in ["assigned_to", "assignee"]:
+            comments.append({"assignee": bug.assigned_to})
+
+    jira_response_comments = []
+    for i, comment in enumerate(comments):
+        logger.debug(
+            "Create comment #%s on Jira issue %s",
+            i + 1,
+            linked_issue_key,
+            extra=log_context.update(operation=Operation.COMMENT).dict(),
+        )
+        jira_response = get_client().issue_add_comment(
+            issue_key=linked_issue_key, comment=json.dumps(comment, indent=4)
+        )
+        jira_response_comments.append(jira_response)
+
+    return jira_response_comments
+
+
+def delete_jira_issue_if_duplicate(context, bug, issue_key):
+    """Rollback the Jira issue creation if there is already a linked Jira issue
+    on the Bugzilla ticket"""
+    # In the time taken to create the Jira issue the bug may have been updated so
+    # re-retrieve it to ensure we have the latest data.
+    jira_key_in_bugzilla = bug.extract_from_see_also()
+    _duplicate_creation_event = (
+        jira_key_in_bugzilla is not None and issue_key != jira_key_in_bugzilla
+    )
+    if not _duplicate_creation_event:
+        return None
+
+    logger.warning(
+        "Delete duplicated Jira issue %s from Bug %s",
+        issue_key,
+        bug.id,
+        extra=context.update(operation=Operation.DELETE).dict(),
+    )
+    jira_response_delete = get_client().delete_issue(issue_id_or_key=issue_key)
+    return jira_response_delete
+
+
+def add_link_to_bugzilla(context, issue_key, bug):
+    """Add link to Bugzilla ticket in Jira issue"""
+    bugzilla_url = f"{settings.bugzilla_base_url}/show_bug.cgi?id={bug.id}"
+    logger.debug(
+        "Link %r on Jira issue %s",
+        bugzilla_url,
+        issue_key,
+        extra=context.update(operation=Operation.LINK).dict(),
+    )
+    icon_url = f"{settings.bugzilla_base_url}/favicon.ico"
+    return get_client().create_or_update_issue_remote_links(
+        issue_key=issue_key,
+        link_url=bugzilla_url,
+        title=bugzilla_url,
+        icon_url=icon_url,
+        icon_title=icon_url,
+    )
+
+
+def clear_assignee(context, issue_key):
+    """Clear the assignee of the specified Jira issue."""
+    logger.debug("Clearing assignee", extra=context.dict())
+    return get_client().update_issue_field(key=issue_key, fields={"assignee": None})
+
+
+def find_jira_user(context, email):
+    """Lookup Jira users, raise an error if not exactly one found."""
+    logger.debug("Find Jira user with email %s", email, extra=context.dict())
+    users = get_client().user_find_by_user_string(query=email)
+    if len(users) != 1:
+        raise ValueError(f"User {email} not found")
+    return users[0]
+
+
+def assign_jira_user(context, issue_key, email):
+    """Set the assignee of the specified Jira issue, raise if fails."""
+    jira_user = find_jira_user(context, email)
+    jira_user_id = jira_user["accountId"]
+    try:
+        # There doesn't appear to be an easy way to verify that
+        # this user can be assigned to this issue, so just try
+        # and do it.
+        return get_client().update_issue_field(
+            key=issue_key,
+            fields={"assignee": {"accountId": jira_user_id}},
+        )
+    except IOError as exc:
+        raise ValueError(
+            f"Could not assign {jira_user_id} to issue {issue_key}"
+        ) from exc
+
+
+def maybe_update_issue_status(context, issue_key, jira_status):
+    """Update the status of the Jira issue or no-op if None."""
+    if jira_status:
+        logger.debug(
+            "Updating Jira status to %s",
+            jira_status,
+            extra=context.dict(),
+        )
+        return get_client().set_issue_status(
+            issue_key,
+            jira_status,
+        )
+
+    logger.debug(
+        "Bug status was not in the status map.",
+        extra=context.update(
+            operation=Operation.IGNORE,
+        ).dict(),
+    )
+    return None
+
+
+def maybe_update_issue_resolution(context, issue_key, jira_resolution):
+    """Update the resolution of the Jira issue or no-op if None."""
+    if jira_resolution:
+        logger.debug(
+            "Updating Jira resolution to %s",
+            jira_resolution,
+            extra=context.dict(),
+        )
+        return get_client().update_issue_field(
+            key=issue_key,
+            fields={"resolution": jira_resolution},
+        )
+
+    logger.debug(
+        "Bug resolution was not in the resolution map.",
+        extra=context.update(
+            operation=Operation.IGNORE,
+        ).dict(),
+    )
+    return None
